+++ bindings/nostr-sdk-ffi/src/relay/status.rs
@@ -16,8 +16,6 @@ pub enum RelayStatus {
     Connected,
     /// Relay disconnected, will retry to connect again
     Disconnected,
-    /// Stop
-    Stopped,
     /// Relay completely disconnected
     Terminated,
 }
@@ -30,7 +28,6 @@ impl From<nostr_sdk::RelayStatus> for RelayStatus {
             nostr_sdk::RelayStatus::Connecting => Self::Connecting,
             nostr_sdk::RelayStatus::Connected => Self::Connected,
             nostr_sdk::RelayStatus::Disconnected => Self::Disconnected,
-            nostr_sdk::RelayStatus::Stopped => Self::Stopped,
             nostr_sdk::RelayStatus::Terminated => Self::Terminated,
         }
     }